Add some methods to gnssDataMap, at the same time, add some global functions to save/load gnssDataMap data to/from a binary file.
Specially, I found that the added following functions is useful for large network data processing, for we can pre-process the data once and save the gnssDataMap object to a binary file, then we can load the preprocessed data any time to process it with different estimation strategy. It save a lot time and easy debuging the software. 

saveGnssDataMap()
loadGnssDataMap()
dumpGnssDataMap()