Small changes to refine and correct SP3 interpolation
...particularly in TabularSatStore, where a developer had modified
the getTableInterval routine; this change is preserved in a renamed
routine, but is not acceptable for SP3EphemerisStore, as noted in
the comments above the new routine declaration. This also involved
the LagrangeInterpolation routines in MiscMath, which have been
made more robust after testing; see comments there.
The change in SP3Data is in comments only, but points out an issue
with SP3 files with epoch lines followed by no data.