Add several new systems to TimeSystem and Time classes.
Add BDS, QZS and GAL time systems. Add abstract classes Week and
WeekSecond to handle all times of the form Week/seconds-of-week,
including GPS. Note GPSWeekZcount is unchanged, and GPSWeek.hpp
exists only for that class. Fix bug in scanf in all time classes in which
time-system was treated as an int when it is a string. Note that
GPSWeekSecond.cpp goes away. Change systemCompass to systemBeiDou
everywhere, and add systemQZSS; do similar for SatID, RinexSatID and
GSatID.