Remove all reference to RinexNav in OrbitEph classes.
This is why there are cast operators in Rinex3NavData; otherwise
there is a circular dependency. Added cast operator to RinexNavData
for backward compatibility, and updated RinexUtilities to use
Rinex3NavData. Fixed Rinex3Nav classes and RinNav to do complete
and correct read and write; tested all systems but with MGEX data.