Build of Core C++ Library No Longer Requires Ext
This means that the core library can be built with the build script without
using the -e option. However, in order for this to be achieved some MAJOR
changes had to be made.

1) The Python hooks that were built by default (not installed without the -p
option) will NOT be made without the -e option. This is because some of those
bindings still required ext.

2) The example code will NOT be built without the -e option.

These changes will be corrected at a later date.