Removed Dependencies on Xv
Xv is a container of position and velocity data. Xvt is a slightly larger
container which contains the same data as Xv but not Xv itself. As Xv is not
widely used, it is to be deprecated. This commit logs the changes for Xv.hpp's
move to deprecate.

The locations where Xv containers are used have been replaced by Xvt. The
includes on those files have been changed to reflect this. Xv.hpp has been moved
to ext/lib/deprecate.