Xv.hpp is now Deprecated
Per request, the Xv.hpp file has been deprecated along with the dependent CNAV
(not to be confused with Cnav) ephemeris tree. This has also required small
changes in the core BrcKeplerOrbit files and the removal of the errant(?)
header in the core HelmertTransform.hpp file.

It will be left up to the GPSTk manager to actually remove the files (and
associated tests).