Updating Rinex3ClockData with the RinexClockData public members
- Imported the public members from RinexClockData to Rinex3ClockData

- Made the duplicate members refrences to each other. If a old script
uses a public member from RinexClockData and modifies the value the
corresponding Rinex3ClockData member is modified as well

- Other members like sat and site from RinexClockData have no
corresponding value in Rinex3ClockData, combined in one value as
name. Created a function called convertTypes that detects if one of
these values were modified, and updates the other values accordingly.