Fixed error within the NumberHeaderRecordsToBeWritten
If this function was called when obsTypeList was empty the
obsTypeList.size() function return 0, which then had -1 added to
it. This caused an unsigned integer rollover, adding an invalid
value to the NumberHeaderRecordsToBeWritten count.

Also, used R2ObsType instead of obsTypeList when the version is
less than 3. R2ObsType list is designed for v.2 Rinex while
obsTypeList designed for v.3 Rinex.