Initial commit of Rinex3Nav_T and all dependent test files
Failing tests:

- hardCodetest fails for writing out a file and comparing it
explicitly to input file. Minor differences between input
and output file, such as additional whitespaces at the end
of each header line and different representation of data (i.e.
.1234E-4 => 1.23E-3) cause test to fail. Input data file needs to
be "cleaned" to match output explicitly if these changes are deemed
insignificant.
- streamWriteTest assumes that by initializing a Rinex3NavStream
object the stream produces a valid header member. Test can either
be modified to write header out to a seperate Rinex3NavHeader
object (which creates a valid header member for the stream), or the
core library code can be modifed to read header of file upon
initialization. Unsure what the streamWrite test is intended to be
testing; comments state RinexEphemerisStore but the methodology is
not apparent. Additionaly, RinexEphemeris store is dependent on
Rinex2 classes.

Commented out two tests from RinexNav_T when porting test to
Rinex3Nav_T, BadHeaderTest and invalidHeaderVersionTest. Both
marked a header as invalid because it's version is 3.33. Rinex3
classes read this version as valid. Unsure if this functionality
is correct or if Rinex3 classes should tag version 3.33 as invalid

View the Rinex3 Testing efforts wiki-page for more detailed
information