Added Rinex version conversion tests to Rinex3Obs_T & Rinex3Nav_T
- Rinex3Nav_T version 2 to 3 and version 3 to 2 tests pass and are
fully automated. However this test is not very rigorous
- Rinex3Obs_T version 2 to 3 and version 3 to 2 tests are NOT
automated. Evaluating the test requires comparing the files output
in the testMesg manually. If these files are accurate conversions
from one version to the other, then an automated test can be
generated by comparing the output of v2->v3 to the input of v3->v2
and vice versa. For now this test is just set to FAIL
- Added a (very) basic data member test for Rinex3Obs_T similar to
what's currently in Rinex3Nav_T