Changes to timeconvert Tests
I had wanted to (and did) add a test for 29-bit zcount. However,
when looking at these tests again I noticed that many of the time
option tests used the same argument list from ANSI time input
due to a copy-paste error. I corrected this.

The new test is failing since timeconvert cannot seem to properly
ingest the 29-bit zcount. And the change in the other tests resulted
in two more errors for julian date and MJD. The results from these
two input types do not match the expected values, nor are they self-
consistent. They claim that the H:M:S of the time entered is 00:01:38
but the SecOfDay is 99. Some truncating/rounding in the CommonTime
conversions may be going wrong.