Removed Improper Use Tests for Apps and Modified PRSolve Tests
After new direction, I have commented out almost all improper use tests for the applications. The
removed tests include bad command line option, missing input file, and invalid input file. The exceptions
to the removal include the invalid input cases for the checktools.

Additionally, the removal of these tests triggered a new failure in PRSolve_Required. After digging it was
found that this was due to a race condition with the creation of log files. The tests have been amended to have a
--log option which points the log file to the same output directory as every other test and provides unique names.