Include SONAME in generated shared library (e.g. libgpstk.so)
Having a valid SONAME has long been a key part of being able to use
shared libraries in a way that allows for dependent applications to
manage ABI compatibility over multiple versions. The SONAME goes into
the library binary (e.g. SONAME field in ELF) and into the filename
of the shared library (e.g. instead of libgpstk.so, you'll now have
libgpstk.so.2.9.7 , which is symlinked as libgpstk.so.2.9 and libgpstk.so.
Among other goodness, this allows multiple versions of the shared library
to be installed simultaneously, and allows the (dynamic) linker to
provide the correct version that each dependent software requires.

See https://en.wikipedia.org/wiki/Soname for more information